separated EmailGeneratorTest from EmailsTest
