+++ src/test/java/teammates/test/cases/logic/EmailsTest.java
@@ -2,7 +2,6 @@ package teammates.test.cases.logic;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.List;
 import java.util.Properties;
 
 import javax.mail.Message;
@@ -13,54 +12,28 @@ import javax.mail.internet.MimeMessage;
 
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import teammates.common.datatransfer.CourseAttributes;
 import teammates.common.datatransfer.FeedbackSessionAttributes;
-import teammates.common.datatransfer.InstructorAttributes;
 import teammates.common.datatransfer.StudentAttributes;
 import teammates.common.datatransfer.UserType;
-import teammates.common.exception.TeammatesException;
 import teammates.common.util.Config;
-import teammates.common.util.Const;
-import teammates.common.util.EmailType;
-import teammates.common.util.StringHelper;
 import teammates.common.util.Templates.EmailTemplates;
 import teammates.common.util.TimeHelper;
-import teammates.logic.core.CoursesLogic;
 import teammates.logic.core.EmailGenerator;
 import teammates.logic.core.Emails;
-import teammates.logic.core.FeedbackSessionsLogic;
-import teammates.logic.core.InstructorsLogic;
 import teammates.logic.core.Sendgrid;
-import teammates.logic.core.StudentsLogic;
 import teammates.test.cases.BaseComponentTestCase;
 import teammates.test.cases.ui.browsertests.SystemErrorEmailReportTest;
-import teammates.test.driver.AssertHelper;
 
 import com.google.appengine.labs.repackaged.org.json.JSONException;
 
 public class EmailsTest extends BaseComponentTestCase {
     
-    private String from; // For MimeMessage testing
-    private String replyTo;
-    
     @BeforeClass
     public static void classSetUp() {
         printTestClassHeader();
-    }
-
-    @BeforeMethod
-    public void caseSetUp() throws IOException {
-        /* For Sendgrid testing */
-        String fromEmail = "Admin@" + Config.getAppId() + ".appspotmail.com";
-        String fromName = "TEAMMATES Admin";
-        replyTo = "teammates@comp.nus.edu.sg";
-        
-        /* For MimeMessage testing */
-        InternetAddress internetAddress = new InternetAddress(fromEmail, fromName);
-        from = internetAddress.toString();
     }
 
     @Test
@@ -94,345 +67,6 @@ public class EmailsTest extends BaseComponentTestCase {
 
     }
     
-    @Test
-    public void testGenerateFeedbackEmailBase() throws IOException, MessagingException {
-
-        FeedbackSessionAttributes fsa = new FeedbackSessionAttributes();
-        fsa.setFeedbackSessionName("Feedback Session Name");
-        fsa.setEndTime(TimeHelper.getDateOffsetToCurrentTime(0));
-
-        CourseAttributes c = new CourseAttributes("course-id", "Course Name");
-
-        StudentAttributes s = new StudentAttributes();
-        s.name = "Student Name";
-        s.key = "skxxxxxxxxxks";
-        s.email = "student@email.tmt";
-
-        @SuppressWarnings("deprecation")
-        InstructorAttributes i = new InstructorAttributes("googleId1", "courseId2", "name", "instructr@email.tmt");
-
-        ______TS("generic template, student yet to join");
-
-        String template = EmailTemplates.USER_FEEDBACK_SESSION;
-        MimeMessage email = new EmailGenerator().generateFeedbackSessionEmailBaseForStudents(
-                c, fsa, s, template, "[Course: %s][Feedback Session: %s]");
-
-        // check receiver
-        assertEquals(s.email, email.getAllRecipients()[0].toString());
-
-        // check sender
-        assertEquals(from, email.getFrom()[0].toString());
-        
-        //check replyTo
-        assertEquals(replyTo, email.getReplyTo()[0].toString());
-
-        // check subject
-        assertEquals("[Course: Course Name][Feedback Session: Feedback Session Name]", email.getSubject());
-
-        // check email body
-        String encryptedKey = StringHelper.encrypt(s.key);
-
-        String submitUrl = Config.getAppUrl(Const.ActionURIs.STUDENT_FEEDBACK_SUBMISSION_EDIT_PAGE)
-                            .withCourseId(c.getId())
-                            .withSessionName(fsa.getFeedbackSessionName())
-                            .withRegistrationKey(encryptedKey)
-                            .withStudentEmail(s.email)
-                            .toAbsoluteString();
-
-        String deadline = TimeHelper.formatTime12H(fsa.getEndTime());
-
-        String emailBody = email.getContent().toString();
-
-        AssertHelper.assertContainsRegex("Hello " + s.name
-                + "{*}${status}{*}" + c.getId() + "{*}" + c.getName() + "{*}"
-                + fsa.getFeedbackSessionName() + "{*}" + deadline + "{*}" + submitUrl + "{*}"
-                + submitUrl, emailBody);
-
-        printEmail(email);
-
-        ______TS("published template, student yet to join");
-
-        template = EmailTemplates.USER_FEEDBACK_SESSION_PUBLISHED;
-        email = new EmailGenerator().generateFeedbackSessionEmailBaseForStudents(c, fsa, s, template, "");
-
-        emailBody = email.getContent().toString();
-
-        assertFalse(emailBody.contains(submitUrl));
-
-        String reportUrl = Config.getAppUrl(Const.ActionURIs.STUDENT_FEEDBACK_RESULTS_PAGE)
-                            .withCourseId(c.getId())
-                            .withSessionName(fsa.getFeedbackSessionName())
-                            .withRegistrationKey(encryptedKey)
-                            .withStudentEmail(s.email)
-                            .toAbsoluteString();
-
-        AssertHelper.assertContainsRegex("Hello " + s.name
-                + "{*}is now open for viewing{*}" + c.getId() + "{*}"
-                + c.getName() + "{*}" + fsa.getFeedbackSessionName() + "{*}" + reportUrl + "{*}"
-                + reportUrl, emailBody);
-
-        printEmail(email);
-
-        ______TS("generic template, student joined");
-
-        s.googleId = "student1id"; // set student id to make him "joined"
-        template = EmailTemplates.USER_FEEDBACK_SESSION;
-
-        email = new EmailGenerator().generateFeedbackSessionEmailBaseForStudents(c, fsa, s, template, "");
-
-        emailBody = email.getContent().toString();
-
-        AssertHelper.assertContainsRegex("Hello " + s.name + "{*}" + c.getId() + "{*}" + c.getName()
-                + "{*}" + fsa.getFeedbackSessionName() + "{*}" + deadline + "{*}" + submitUrl + "{*}"
-                + submitUrl, emailBody);
-
-        printEmail(email);
-
-        ______TS("published template, student joined");
-
-        template = EmailTemplates.USER_FEEDBACK_SESSION_PUBLISHED;
-        email = new EmailGenerator().generateFeedbackSessionEmailBaseForStudents(c, fsa, s, template, "");
-
-        emailBody = email.getContent().toString();
-
-        AssertHelper.assertContainsRegex("Hello " + s.name
-                + "{*}is now open for viewing{*}" + c.getId() + "{*}" + c.getName()
-                + "{*}" + fsa.getFeedbackSessionName() + "{*}" + reportUrl + "{*}" + reportUrl,
-                emailBody);
-
-        printEmail(email);
-        
-        ______TS("generic template, sent to instructors");
-        
-        template = EmailTemplates.USER_FEEDBACK_SESSION;
-        email = new EmailGenerator().generateFeedbackSessionEmailBaseForInstructors(c, fsa, i, template, "");
-
-        emailBody = email.getContent().toString();
-
-        assertFalse(emailBody.contains("${joinFragment}"));
-        
-        AssertHelper.assertContainsRegex("Hello " + i.name + "{*}"
-                + "The email below has been sent to students of course: " + c.getId()
-                + "{*}" + c.getId() + "{*}" + c.getName()
-                + "{*}" + fsa.getFeedbackSessionName() + "{*}" + deadline
-                + "{*}{The student's unique submission url appears here}"
-                + "{*}{The student's unique submission url appears here}",
-                emailBody);
-
-        printEmail(email);
-        
-        ______TS("published template, sent to instructors");
-        
-        template = EmailTemplates.USER_FEEDBACK_SESSION_PUBLISHED;
-        email = new EmailGenerator().generateFeedbackSessionEmailBaseForInstructors(c, fsa, i, template, "");
-
-        emailBody = email.getContent().toString();
-
-        assertFalse(emailBody.contains("${joinFragment}"));
-        
-        AssertHelper.assertContainsRegex("Hello " + i.name + "{*}"
-                + "The email below has been sent to students of course: " + c.getId()
-                + "{*}is now open for viewing{*}" + c.getId() + "{*}" + c.getName()
-                + "{*}" + fsa.getFeedbackSessionName()
-                + "{*}{The student's unique results url appears here}"
-                + "{*}{The student's unique results url appears here}",
-                emailBody);
-
-        printEmail(email);
-        
-    }
-
-    @Test
-    public void testGenerateStudentCourseJoinEmail() throws IOException, MessagingException {
-
-        CourseAttributes c = new CourseAttributes("course-id", "Course Name");
-
-        StudentAttributes s = new StudentAttributes();
-        s.name = "Student Name";
-        s.key = "skxxxxxxxxxks";
-        s.email = "student@email.tmt";
-
-        MimeMessage email = new EmailGenerator().generateStudentCourseJoinEmail(c, s);
-
-        // check receiver
-        assertEquals(s.email, email.getAllRecipients()[0].toString());
-
-        // check sender
-        assertEquals(from, email.getFrom()[0].toString());
-        
-        //check replyTo
-        assertEquals(replyTo, email.getReplyTo()[0].toString());
-        
-        // check subject
-        assertEquals(
-                "TEAMMATES: Invitation to join course [Course Name][Course ID: course-id]",
-                email.getSubject());
-
-        // check email body
-        String joinUrl = Config.getAppUrl(s.getRegistrationUrl()).toAbsoluteString();
-        String emailBody = email.getContent().toString();
-
-        AssertHelper.assertContainsRegex("Hello " + s.name + "{*}course <i>" + c.getName()
-                + "{*}" + joinUrl + "{*}" + joinUrl + "{*}", emailBody);
-        
-        assertFalse(emailBody.contains("$"));
-
-        printEmail(email);
-    }
-
-    private void printEmail(MimeMessage email) throws MessagingException,
-            IOException {
-        print("Here's the generated email (for your eyeballing pleasure):");
-        print(".............[Start of email]..............");
-        print("Subject: " + email.getSubject());
-        print("Body:");
-        print(email.getContent().toString());
-        print(".............[End of email]................");
-    }
-    
-    @Test
-    public void testGenerateFeedbackSessionEmails() throws Exception {
-
-        removeAndRestoreTypicalDataInDatastore();
-        
-        StudentsLogic studentsLogic = StudentsLogic.inst();
-        InstructorsLogic instructorsLogic = InstructorsLogic.inst();
-        CoursesLogic coursesLogic = CoursesLogic.inst();
-        FeedbackSessionsLogic fsLogic = FeedbackSessionsLogic.inst();
-        
-        FeedbackSessionAttributes fsa = fsLogic.getFeedbackSession("First feedback session", "idOfTypicalCourse1");
-        
-        CourseAttributes c = coursesLogic.getCourse(fsa.getCourseId());
-        
-        List<StudentAttributes> students = studentsLogic.getStudentsForCourse(fsa.getCourseId());
-        List<InstructorAttributes> instructors = instructorsLogic.getInstructorsForCourse(fsa.getCourseId());
-        
-        StudentAttributes s1 = new StudentAttributes();
-        s1.email = "student5InCourse1@gmail.tmt";
-
-        StudentAttributes s2 = new StudentAttributes();
-        s2.email = "student2InCourse1@gmail.tmt";
-        
-        StudentAttributes s3 = new StudentAttributes();
-        s3.email = "student3InCourse1@gmail.tmt";
-
-        @SuppressWarnings("deprecation")
-        InstructorAttributes i1 = new InstructorAttributes("googleId", "courseId1", "name", "instructor1@course1.tmt");
-
-        @SuppressWarnings("deprecation")
-        InstructorAttributes i2 = new InstructorAttributes("googleId", "courseId1", "name", "instructor2@course1.tmt");
-        
-        ______TS("feedback session opening emails");
-
-        List<MimeMessage> emails = new EmailGenerator().generateFeedbackSessionOpeningEmails(fsa);
-        assertEquals(10, emails.size());
-
-        String subject = String.format(EmailType.FEEDBACK_OPENING.getSubject(), c.getName(), fsa.getFeedbackSessionName());
-        String status = "is now open";
-        verifyEmail(s1, emails.get(0), subject, status);
-        verifyEmail(s2, emails.get(1), subject, status);
-        
-        ______TS("feedback session reminders");
-
-        emails = new EmailGenerator().generateFeedbackSessionReminderEmails(c, fsa, students, instructors, instructors);
-        assertEquals(15, emails.size());
-
-        subject = String.format(EmailType.FEEDBACK_SESSION_REMINDER.getSubject(), c.getName(), fsa.getFeedbackSessionName());
-        status = "is still open for submissions";
-        verifyEmail(i1, emails.get(1), subject, status);
-        verifyEmail(i2, emails.get(2), subject, status);
-        
-        ______TS("feedback session closing alerts");
-
-        emails = new EmailGenerator().generateFeedbackSessionClosingEmails(fsa);
-        assertEquals(8, emails.size());
-
-        subject = String.format(EmailType.FEEDBACK_CLOSING.getSubject(), c.getName(), fsa.getFeedbackSessionName());
-        status = "is closing soon";
-        verifyEmail(s1, emails.get(0), subject, status);
-        verifyEmail(s3, emails.get(1), subject, status);
-        String ignoreEmailMsg = "You may ignore this email if you have already submitted feedback.";
-        verifyEmail(s1, emails.get(0), subject, ignoreEmailMsg);
-        verifyEmail(s3, emails.get(1), subject, ignoreEmailMsg);
-
-        ______TS("feedback session published alerts");
-
-        emails = new EmailGenerator().generateFeedbackSessionPublishedEmails(fsa);
-        assertEquals(10, emails.size());
-
-        subject = String.format(EmailType.FEEDBACK_PUBLISHED.getSubject(), c.getName(), fsa.getFeedbackSessionName());
-        status = "The feedback responses for the following feedback session is now open for viewing.";
-        verifyEmail(s1, emails.get(0), subject, status);
-        verifyEmail(s2, emails.get(1), subject, status);
-    }
-    
-    @Test
-    public void testSystemCrashReportEmailContent() throws IOException,
-            MessagingException {
-
-        
-        AssertionError error = new AssertionError("invalid parameter");
-        StackTraceElement s1 = new StackTraceElement(
-                SystemErrorEmailReportTest.class.getName(),
-                "testSystemCrashReportEmailContent",
-                "SystemErrorEmailReportTest.java",
-                89);
-        error.setStackTrace(new StackTraceElement[] {s1});
-        String stackTrace = TeammatesException.toStringWithStackTrace(error);
-        String requestMethod = "GET";
-        String requestUserAgent = "user-agent";
-        String requestPath = "/page/studentHome";
-        String requestUrl = "/page/studentHome/";
-        String requestParam = "{}";
-        UserType userType = new UserType("Not logged in");
-
-        MimeMessage email =
-                new EmailGenerator().generateSystemErrorEmail(requestMethod, requestUserAgent, requestPath,
-                                                              requestUrl, requestParam, userType, error);
-
-        // check receiver
-        String recipient = Config.SUPPORT_EMAIL;
-        assertEquals(recipient, email.getAllRecipients()[0].toString());
-
-        // check sender
-        assertEquals(from, email.getFrom()[0].toString());
-        
-            
-        // check email body
-        String emailBody = email.getContent().toString();
-        AssertHelper.assertContainsRegex(
-                "<b>Error Message</b><br/><pre><code>" + error.getMessage()
-                + "</code></pre>"
-                + "<br/><b>Actual user</b>" + "Not logged in"
-                + "<br/><b>Request Method</b>" + requestMethod
-                + "<br/><b>User Agent</b>" + requestUserAgent
-                + "<br/><b>Request Url</b>" + requestUrl
-                + "<br/><b>Request Path</b>" + requestPath
-                + "<br/><b>Request Parameters</b>" + requestParam
-                + "<br/><b>Stack Trace</b><pre><code>" + stackTrace + "</code></pre>",
-                emailBody);
-    }
-
-    private void verifyEmail(StudentAttributes s, MimeMessage email,
-            String subject, String textInEmail) throws MessagingException,
-            IOException {
-        assertEquals(s.email, email.getAllRecipients()[0].toString());
-        assertEquals(subject, email.getSubject());
-        String emailBody = email.getContent().toString();
-        assertTrue(emailBody.contains(textInEmail));
-        assertFalse(emailBody.contains("$"));
-    }
-    
-    private void verifyEmail(InstructorAttributes i, MimeMessage email,
-            String subject, String textInEmail) throws MessagingException,
-            IOException {
-        assertEquals(i.email, email.getAllRecipients()[0].toString());
-        assertEquals(subject, email.getSubject());
-        String emailBody = email.getContent().toString();
-        assertTrue(emailBody.contains(textInEmail));
-        assertFalse(emailBody.contains("$"));
-    }
-
     @Test
     public void testNoExceptionThrownWhenNoMessagesToSend() {
         new Emails().sendEmails(new ArrayList<MimeMessage>());
@@ -505,7 +139,7 @@ public class EmailsTest extends BaseComponentTestCase {
         assertEquals(email.getContent().toString(), sendgridEmail.getHtml());
     }
 
-    @AfterClass()
+    @AfterClass
     public static void classTearDown() {
         printTestClassFooter();
     }
